requirements.txt
pyodbc
pyserial
websockets
python-dotenv
pytest